pyproject.toml
[project]
name = "civicpulse"
version = "0.1.0"
description = "Civic engagement AI agent for Town of Babylon, NY"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0",
    "httpx>=0.27",
    "fastapi>=0.115",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pdfplumber>=0.11",
    "sqlite-utils>=3.37",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "click>=8.1",
    "python-slugify>=8.0",
    "python-dotenv>=1.0",
    "uvicorn>=0.32",
    "markitdown>=0.1.5",
    "youtube-transcript-api>=1.2.4",
    "google-api-python-client>=2.194.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.67",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-pyyaml",
    "types-beautifulsoup4",
]

[project.scripts]
civicpulse-scrape = "civicpulse.scraper.cli:scrape"
civicpulse-query = "civicpulse.scraper.cli:query"
civicpulse-import = "civicpulse.scraper.cli:import_documents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/civicpulse"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
